apps/python/flask/flask-app-2-es/requirements.txt
Flask
elasticsearch
python-dotenv
